+++ modules/canvaskit/externs.js
@@ -77,13 +77,9 @@ var CanvasKit = {
   // in the pre-js file)
   _computeTonalColors: function() {},
   _MakeImage: function() {},
-  _MakeLinearGradientShader: function() {},
-  _MakeRadialGradientShader: function() {},
-  _MakeSweepGradientShader: function() {},
   _MakeManagedAnimation: function() {},
   _MakeParticles: function() {},
   _MakePicture: function() {},
-  _MakeTwoPointConicalGradientShader: function() {},
   _decodeAnimatedImage: function() {},
   _decodeImage: function() {},
   _drawShapedText: function() {},
@@ -604,17 +600,25 @@ var CanvasKit = {
   },
 
   Shader: {
+    // Deprecated names
     Blend: function() {},
     Color: function() {},
-    Empty: function() {},
     Lerp: function() {},
+    // public API (from JS)
+    MakeBlend: function() {},
+    MakeColor: function() {},
+    MakeLerp: function() {},
     MakeLinearGradient: function() {},
     MakeRadialGradient: function() {},
     MakeTwoPointConicalGradient: function() {},
     MakeSweepGradient: function() {},
 
     // private API (from C++ bindings)
-    _Color: function() {},
+    _MakeColor: function() {},
+    _MakeLinearGradient: function() {},
+    _MakeRadialGradient: function() {},
+    _MakeSweepGradient: function() {},
+    _MakeTwoPointConicalGradient: function() {},
   },
 
   Surface: {